social_media_api/requirements.txt
asgiref==3.10.0
dj-database-url==3.0.1
Django==5.2.7
django-csp==4.0
django-filter==25.2
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
